Changed cart items to be stored in session instead of the database.

+++ Shop.Infrastructure/Repositories/CartItemRepository.cs
@@ -1,12 +1,8 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.AspNetCore.Http;
+using Microsoft.EntityFrameworkCore;
 using Shop.Domain.Entities;
 using Shop.Domain.Interfaces;
 using Shop.Infrastructure.Persistence;
-using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace Shop.Infrastructure.Repositories
@@ -14,27 +10,50 @@ namespace Shop.Infrastructure.Repositories
     public class CartItemRepository : ICartItemRepository
     {
         private readonly ShopDbContext _dbContext;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public CartItemRepository(ShopDbContext dbContext)
+        public CartItemRepository(ShopDbContext dbContext, IHttpContextAccessor httpContextAccessor)
         {
             _dbContext = dbContext;
+            _httpContextAccessor = httpContextAccessor;
         }
+
         public async Task Create(CartItem cartItem)
         {
+            if (cartItem == null)
+            {
+                throw new ArgumentNullException(nameof(cartItem));
+            }
+
             _dbContext.CartItems.Add(cartItem);
             await _dbContext.SaveChangesAsync();
         }
+
         public async Task<CartItem> GetCartItem(int cartItemId)
         {
-            var cartItem = await _dbContext.CartItems.FirstOrDefaultAsync(c => c.Id == cartItemId);
+            if (cartItemId <= 0)
+            {
+                throw new ArgumentException("Invalid cartItemId");
+            }
+
+            var cartItem = await _dbContext.CartItems
+                .Include(c => c.Item)
+                .FirstOrDefaultAsync(c => c.Id == cartItemId);
 
             return cartItem;
         }
+
         public async Task Delete(CartItem cartItem)
         {
+            if (cartItem == null)
+            {
+                throw new ArgumentNullException(nameof(cartItem));
+            }
+
             _dbContext.CartItems.Remove(cartItem);
             await _dbContext.SaveChangesAsync();
         }
+
         public async Task Commit()
         {
             await _dbContext.SaveChangesAsync();
@@ -42,11 +61,21 @@ namespace Shop.Infrastructure.Repositories
 
         public async Task UpdateCartItemQuantity(CartItem cartItem, int quantity)
         {
+            if (cartItem == null)
+            {
+                throw new ArgumentNullException(nameof(cartItem));
+            }
+
             cartItem.Quantity = quantity;
             await _dbContext.SaveChangesAsync();
         }
+
         public async Task UpdateCartItem(CartItem cartItem)
-        { 
+        {
+            if (cartItem == null)
+            {
+                throw new ArgumentNullException(nameof(cartItem));
+            }
 
             _dbContext.CartItems.Update(cartItem);
             await _dbContext.SaveChangesAsync();
